@@ -1,6 +1,6 @@
 {
   "name": "howz.dev",
-  "version": "0.0.6",
+  "version": "0.1.0",
   "description": "💠 Personal website and blog made using Next.js, TypeScript, Tailwind CSS and Notion",
   "author": "howznguyen",
   "license": "ISC",
@@ -9,19 +9,18 @@
     "url": "https://github.com/howznguyen/howz.dev.git"
   },
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
-    "postbuild": "next-sitemap",
     "start": "next start",
     "lint": "next lint",
-    "prebuild": "next build && next-sitemap",
+    "prebuild": "next build",
     "start:all": "next build && next start"
   },
   "dependencies": {
     "@giscus/react": "^2.2.6",
     "@next/bundle-analyzer": "^13.2.4",
-    "@next/font": "^13.1.6",
     "@notionhq/client": "^2.2.3",
+    "@types/lodash": "^4.17.20",
     "@types/node": "18.11.18",
     "@types/react": "18.0.27",
     "@types/react-dom": "18.0.10",
@@ -29,13 +28,20 @@
     "eslint": "8.32.0",
     "eslint-config-next": "13.1.6",
     "got": "^12.5.3",
+    "lodash": "^4.17.21",
+    "media-chrome": "^4.13.1",
     "moment": "^2.29.4",
-    "next": "13.1.6",
+    "next": "15.5.3",
     "next-seo": "^5.15.0",
     "next-sitemap": "^3.1.52",
     "next-themes": "^0.2.1",
     "nextjs-google-analytics": "^2.3.0",
+    "notion-client": "^7.6.0",
     "notion-to-md": "^2.5.5",
+    "notion-types": "^7.6.0",
+    "ofetch": "^1.4.1",
+    "open-graph-scraper": "^6.10.0",
+    "p-memoize": "^8.0.0",
     "prismjs": "^1.29.0",
     "react": "18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
@@ -46,9 +52,8 @@
     "react-medium-image-zoom": "^5.1.2",
     "react-nl2br": "^1.0.4",
     "react-player": "^2.12.0",
-    "react-scrollspy": "^3.4.3",
     "react-social-media-embed": "^2.3.4",
-    "reading-time": "^1.5.0",
+    "reading-time-estimator": "^1.14.0",
     "remove-accents": "^0.4.4",
     "sharp": "^0.31.3",
     "typescript": "4.9.4"
@@ -58,7 +63,6 @@
     "@tailwindcss/typography": "^0.5.9",
     "@types/prismjs": "^1.26.0",
     "@types/react-copy-to-clipboard": "^5.0.4",
-    "@types/react-scrollspy": "^3.3.5",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.2.4"
