@@ -1,21 +1,21 @@
+"use client";
+
 import { useTheme } from "next-themes";
 import Button from "./Button";
 import Icon from "./Icon";
-import { useTrans } from "@/lib";
-
+import header from "@/datas/header";
 
 interface ThemeSwitcherProps {}
 
 const ThemeSwitcher = ({}: ThemeSwitcherProps) => {
   const { theme, setTheme } = useTheme();
-  const trans = useTrans();
-  
+
   return (
     <Button
       onClick={() => setTheme(theme === "dark" ? "light" : "dark")}
-      title={ trans.header.theme_switcher }
+      title={header.theme_switcher}
     >
-      {theme === "dark" ? <Icon icon="BiMoon" />  : <Icon icon="BiSun" />}
+      {theme === "dark" ? <Icon icon="BiMoon" /> : <Icon icon="BiSun" />}
     </Button>
   );
 };
