@@ -1,26 +1,9 @@
-export { default as Button } from './Button';
-export { default as Link } from './Link';
-export { default as NavItem } from './NavItem';
-export { default as Logo } from './Logo';
-export { default as ThemeSwitcher } from './ThemeSwitcher';
-export { default as DateTime } from './DateTime';
-export { default as Loading} from './Loading';
-export { default as Icon} from './Icon';
-export { default as LangugeSwither} from './LangugeSwither';
-
-// Notion Content
-export { default as Heading } from './Heading';
-export { default as Blockquote } from './Blockquote';
-export { default as Image } from './Image';
-export { default as UnorderedList } from './UnorderedList';
-export { default as OrderedList } from './OrderedList';
-export { default as CodeBlock } from './CodeBlock';
-export { default as Callout } from './Callout';
-export { default as Paragraph } from './Paragraph';
-export { default as Span } from './Span';
-export { default as Toggle } from './Toggle';
-export { default as Bookmark } from './Bookmark';
-export { default as Table } from './Table';
-export { default as Tag} from './Tag';
-export { default as Embed} from './Embed';
-export { default as Video} from './Video';
+export { default as Button } from "./Button";
+export { default as Link } from "./Link";
+export { default as NavItem } from "./NavItem";
+export { default as Logo } from "./Logo";
+export { default as ThemeSwitcher } from "./ThemeSwitcher";
+export { default as DateTime } from "./DateTime";
+export { default as Loading } from "./Loading";
+export { default as Icon } from "./Icon";
+export { default as Tag } from "./Tag";
