@@ -1,7 +1,7 @@
-export { default as Navigation } from './Navigation';
-export { default as PostCard } from './PostCard';
-export { default as CommentSection } from './CommentSection';
-export { default as TableOfContents } from './TableOfContents';
-export { default as PostList } from './PostList';
-export { default as NotionRender } from './NotionRender';
-export { default as IntroCard } from './IntroCard';
+export { default as Navigation } from "./Navigation";
+export { default as PostCard } from "./PostCard";
+export { default as CommentSection } from "./CommentSection";
+export { default as TableOfContents } from "./TableOfContents";
+export { default as PostList } from "./PostList";
+export { default as IntroCard } from "./IntroCard";
+export { default as Views } from "./Views";
