@@ -1,27 +1,51 @@
-import { BASE_URL } from '@/lib/env';
+import { BASE_URL } from "@/lib/env";
 
 const Route = {
-    index: (includeDomain = false) => `${includeDomain ? BASE_URL : ''}/`,
-    post: (slug: string, includeDomain = false) => `${includeDomain ? BASE_URL : ''}/post/${slug}`,
-    blog: (includeDomain = false) => `${includeDomain ? BASE_URL : ''}/blog`,
-    author: (slug: string, includeDomain = false) => `${includeDomain ? BASE_URL : ''}/author/${slug}`,
-    tag: {
-        index: (includeDomain = false) => `${includeDomain ? BASE_URL : ''}/tag`,
-        get: (slug: string, includeDomain = false) => `${includeDomain ? BASE_URL : ''}/tag/${slug}`,
+  index: (includeDomain = false) => {
+    const base = includeDomain ? BASE_URL : "";
+    return `${base}/`;
+  },
+  post: (slug: string, includeDomain = false) => {
+    const base = includeDomain ? BASE_URL : "";
+    return `${base}/post/${slug}`;
+  },
+  blog: (includeDomain = false) => {
+    const base = includeDomain ? BASE_URL : "";
+    return `${base}/blog`;
+  },
+  author: (slug: string, includeDomain = false) => {
+    const base = includeDomain ? BASE_URL : "";
+    return `${base}/author/${slug}`;
+  },
+  tag: {
+    index: (includeDomain = false) => {
+      const base = includeDomain ? BASE_URL : "";
+      return `${base}/tag`;
     },
-
-    // API
-    api: {
-        post: {
-            updateViews : (slug: string, language: string, includeDomain = false) => `${includeDomain ? BASE_URL : ''}/api/post/views?slug=${slug}&language=${language}`
-        }
+    get: (slug: string, includeDomain = false) => {
+      const base = includeDomain ? BASE_URL : "";
+      return `${base}/tag/${slug}`;
     },
+  },
 
-    // Image
-    image: (file: string, includeDomain = false) => `${includeDomain ? BASE_URL : ''}/assets/images/${file}`,
-    defaultCover: (includeDomain = false) => `${includeDomain ? BASE_URL : ''}/assets/images/og.png`,
-    defaultLogo: (includeDomain = false) => `${includeDomain ? BASE_URL : ''}/assets/images/logo.png`,
+  // RSS
+  rss: (includeDomain = false) => {
+    const base = includeDomain ? BASE_URL : "";
+    return `${base}/rss.xml`;
+  },
+
+  // API
+  api: {
+    // Removed post views API
+  },
+
+  // Image
+  image: (file: string, includeDomain = false) =>
+    `${includeDomain ? BASE_URL : ""}/assets/images/${file}`,
+  defaultCover: (includeDomain = false) =>
+    `${includeDomain ? BASE_URL : ""}/assets/images/og.png`,
+  defaultLogo: (includeDomain = false) =>
+    `${includeDomain ? BASE_URL : ""}/assets/images/logo.png`,
 };
-
 
 export default Route;
